Move transaction-related commands to eth_txs.go
